Sawa-Coco-Bot/requirements.txt
python-telegram-bot==20.7
openai==1.3.8
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
